委员详情细节优化

+++ ELA/Wallet/CRCommtieeViewController/Model/CouncilAndSecretariatDetail/ELAInformationDetail.h
@@ -29,8 +29,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface ELAInformationDetail : ELABaseModel
 
-@property (nonatomic, assign) double impeachmentThroughVotes;
-@property (nonatomic, assign) double impeachmentVotes;
+@property (nonatomic, assign) long impeachmentThroughVotes;
+@property (nonatomic, assign) long impeachmentVotes;
 @property (nonatomic, strong) NSString *did;
 @property (nonatomic, strong) NSString *cid;
 @property (nonatomic, strong) NSString *didName;
